Refactored prefixes, removed entries from spelling wordlist
