+++ airflow/providers/databricks/sensors/partition.py
@@ -26,7 +26,7 @@ from typing import Sequence
 from databricks.sql.utils import ParamEscaper
 
 from airflow.exceptions import AirflowException
-from airflow.providers.databricks.sensors.databricks_sql import DatabricksSqlSensor
+from airflow.providers.databricks.sensors.sql import DatabricksSqlSensor
 from airflow.utils.context import Context
