+++ docs/spelling_wordlist.txt
@@ -339,9 +339,6 @@ dask
 dat
 Databricks
 databricks
-DatabricksPartitionSensor
-DatabricksSqlSensor
-DatabricksTableChangesSensor
 datacenter
 dataclass
 Datadog
